backend/requirements/base.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==1.4.53
databases[sqlite]==0.8.0
alembic==1.13.1
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Telegram
telethon==1.34.0

# Scheduling
apscheduler==3.10.4

# HTTP
python-multipart==0.0.6

